chore: if the fixture is in the right file, it works
