+++ tests/unit/conftest.py
@@ -1,14 +1,11 @@
 # Copyright 2024 Canonical Ltd.
 # See LICENSE file for licensing details.
 
-"""Fixtures for maubotunit tests."""
+"""Fixtures for maubot unit tests."""
 
-import json
 
 import pytest
-import pytest_asyncio
 from ops.testing import Harness
-from pytest_operator.plugin import OpsTest
 
 from charm import MaubotCharm
 
@@ -40,27 +37,3 @@ admins:
     (root / "data" / "config.yaml").write_text(yaml_content)
     yield harness
     harness.cleanup()
-
-
-@pytest_asyncio.fixture(scope="function", name="get_unit_ips")
-async def fixture_get_unit_ips(ops_test: OpsTest):
-    """Return an async function to retrieve unit ip addresses of a certain application."""
-
-    async def get_unit_ips(application_name: str):
-        """Retrieve unit ip addresses of a certain application.
-
-        Args:
-            application_name: application to get the ip address.
-
-        Returns:
-            A list containing unit ip addresses.
-        """
-        _, status, _ = await ops_test.juju("status", "--format", "json")
-        status = json.loads(status)
-        units = status["applications"][application_name]["units"]
-        return tuple(
-            unit_status["address"]
-            for _, unit_status in sorted(units.items(), key=lambda kv: int(kv[0].split("/")[-1]))
-        )
-
-    return get_unit_ips
